Switch to reputable, pure Python libraries.
To offer stronger guarantees that the code will run in other
implementations of Python, and to avoid pesky compilation steps on
lesser platforms, we switch to select pure Python libraries after a
round of evaluation of the available options. The results of this
switch, compared with both the prior Pydantic implementation and the
original dictionary-based implementation by mhdadk, and in terms of
the total dependency count and peak memory usage when run on a dummy
account (ACCOUNT_ID = 8), are as follows:
* dict: 5 dependencies; 41 MB
* Pydantic: 21 dependencies; 60 MB
* attrs: 10 dependencies; 48 MB

All typing issues are fixed: the suppressed warnings are either false
positives or ignorable. Detection of self-answered questions has been
restored to past performance before the output format redesign. There
is cleaner separation of concerns and more additions to CLI options in
anticipation of certain use cases.